+++ project/src/components/login-form/login-form.tsx
@@ -1,0 +1,74 @@
+import { useRef, FormEvent, useCallback } from 'react';
+import { AuthData } from '../../types/auth-data';
+import { useAppDispatch } from '../../hooks';
+import { loginAction } from '../../store/api-actions';
+import { toast } from 'react-toastify';
+import {validateLoginForm} from '../../utils/utils';
+
+
+export default function LoginForm(): JSX.Element {
+  const dispatch = useAppDispatch();
+  const loginRef = useRef<HTMLInputElement | null>(null);
+  const passwordRef = useRef<HTMLInputElement | null>(null);
+
+  const onSubmit = useCallback((authData: AuthData) => {
+    dispatch(loginAction(authData));
+  },[dispatch]);
+
+  const handleSubmit = useCallback((evt: FormEvent<HTMLFormElement>) => {
+    evt.preventDefault();
+
+    if (loginRef.current !== null
+      && passwordRef.current !== null
+      && validateLoginForm(loginRef.current, passwordRef.current )
+    ) {
+      onSubmit({
+        login: loginRef.current.value,
+        password: passwordRef.current.value,
+      });
+    } else {
+      toast.warn('Invalid Email or password. The password must include of at least one letter and a number.');
+    }
+  },[onSubmit]);
+
+  return (
+    <section className="login">
+      <h1 className="login__title">Sign in</h1>
+      <form
+        className="login__form form"
+        action="#"
+        method="post"
+        onSubmit={handleSubmit}
+      >
+        <div className="login__input-wrapper form__input-wrapper">
+          <label className="visually-hidden">E-mail</label>
+          <input
+            ref={loginRef}
+            className="login__input form__input"
+            type="email"
+            name="email"
+            placeholder="Email"
+            required
+          />
+        </div>
+        <div className="login__input-wrapper form__input-wrapper">
+          <label className="visually-hidden">Password</label>
+          <input
+            ref={passwordRef}
+            className="login__input form__input"
+            type="password"
+            name="password"
+            placeholder="Password"
+            required
+          />
+        </div>
+        <button
+          className="login__submit form__submit button"
+          type="submit"
+        >
+          Sign in
+        </button>
+      </form>
+    </section>
+  );
+}
