+++ project/src/pages/login-screen/login-screen.tsx
@@ -1,35 +1,18 @@
-import { FormEvent, useCallback, useRef } from 'react';
+import { Link } from 'react-router-dom';
 import Header from '../../components/header/header';
+import LoginForm from '../../components/login-form/login-form';
 import { AppRoute, AuthorizationStatus } from '../../const';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { redirectToRoute } from '../../store/action';
-import { loginAction } from '../../store/api-actions';
 import { getAuthorizationStatus } from '../../store/user-process/selector';
-import { AuthData } from '../../types/auth-data';
 
 
 function LoginScreen(): JSX.Element {
 
-  const loginRef = useRef<HTMLInputElement | null>(null);
-  const passwordRef = useRef<HTMLInputElement | null>(null);
+
   const authorization = useAppSelector(getAuthorizationStatus);
 
   const dispatch = useAppDispatch();
-
-  const onSubmit = useCallback((authData: AuthData) => {
-    dispatch(loginAction(authData));
-  },[dispatch]);
-
-  const handleSubmit = useCallback((evt: FormEvent<HTMLFormElement>) => {
-    evt.preventDefault();
-
-    if (loginRef.current !== null && passwordRef.current !== null) {
-      onSubmit({
-        login: loginRef.current.value,
-        password: passwordRef.current.value,
-      });
-    }
-  },[onSubmit]);
 
   if (authorization === AuthorizationStatus.Auth) {
     dispatch(redirectToRoute(AppRoute.Main));
@@ -41,48 +24,12 @@ function LoginScreen(): JSX.Element {
 
       <main className="page__main page__main--login">
         <div className="page__login-container container">
-          <section className="login">
-            <h1 className="login__title">Sign in</h1>
-            <form
-              className="login__form form"
-              action="#"
-              method="post"
-              onSubmit ={handleSubmit}
-            >
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">E-mail</label>
-                <input
-                  ref={loginRef}
-                  className="login__input form__input"
-                  type="email"
-                  name="email"
-                  placeholder="Email"
-                  id= 'email'
-                  required
-                />
-              </div>
-              <div className="login__input-wrapper form__input-wrapper">
-                <label className="visually-hidden">Password</label>
-                <input
-                  ref={passwordRef}
-                  className="login__input form__input"
-                  type="password"
-                  name="password"
-                  placeholder="Password"
-                  id = 'password'
-                  required
-                />
-              </div>
-              <button className="login__submit form__submit button" type="submit">
-                Sign in
-              </button>
-            </form>
-          </section>
+          <LoginForm />
           <section className="locations locations--login locations--current">
             <div className="locations__item">
-              <a className="locations__item-link" href="/#">
+              <Link to={AppRoute.Main }className="locations__item-link" >
                 <span>Amsterdam</span>
-              </a>
+              </Link>
             </div>
           </section>
         </div>
