+++ project/src/utils/utils.tsx
@@ -105,6 +105,14 @@ const getRandomId = (min: number, max: number): string => {
 };
 
 
+const validateLoginForm = (email: HTMLInputElement, password: HTMLInputElement) => {
+  const validPassword = /^.*(?=.{1,})(?=.*[a-zA-Z])(?=.*\d)/;
+  const validLogin = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+
+  return !!(password.value.match(validPassword) && email.value.match(validLogin));
+};
+
+
 export {
   sortByCity,
   ratingPercentage,
@@ -118,5 +126,6 @@ export {
   changeMockOfferFavoriteStatus,
   changeOneItem,
   updateFavorites,
-  getRandomId
+  getRandomId,
+  validateLoginForm
 };
